changed to double

+++ example2/Program.cs
@@ -1,7 +1,7 @@
 ﻿Console.WriteLine("Введите число");
-int num1 = Convert.ToInt32(Console.ReadLine());
+double num1 = Convert.ToDouble(Console.ReadLine());
 System.Console.WriteLine("Введите второе число");
-int num2 = Convert.ToInt32(Console.ReadLine());
+double num2 = Convert.ToDouble(Console.ReadLine());
 
 if (num1 > num2)
 {
